Componentize expandable content
